Broke input modifiers into more modules
