+++ armi/cases/inputModifiers/__init__.py
@@ -1,0 +1,9 @@
+"""
+Code that changes input files and writes them back out. Useful for parameter sweeps.
+
+See Also
+--------
+armi.reactor.converters
+    Code that changes reactor objects at runtime. These often take longer to run than
+    these but can be used in the middle of ARMI analyses.
+"""
