+++ armi/cases/suiteBuilder.py
@@ -30,8 +30,6 @@ import copy
 import os
 import random
 
-from armi.reactor import flags
-from armi.reactor.components import component
 from armi.cases import suite
 
 
@@ -296,251 +294,3 @@ class SeparateEffectsSuiteBuilder(SuiteBuilder):
         SuiteBuilder.addDegreeOfFreedom
         """
         self.modifierSets.extend((modifier,) for modifier in inputModifiers)
-
-
-class InputModifier(object):
-    """
-    Object that modifies input definitions in some well-defined way.
-
-    (This class is abstract.)
-
-    Subclasses must implement a ``__call__`` method accepting a ``CaseSettings``,
-    ``Blueprints``, and ``SystemLayoutInput``.
-
-    The class attribute ``FAIL_IF_AFTER`` should be a tuple defining what, if any,
-    modifications this should fail if performed after. For example, one should not
-    adjust the smear density (a function of Cladding ID) before adjusting the Cladding
-    ID.
-
-    Some subclasses are provided, but you are expected to make your own design-specific
-    modifiers in most cases.
-    """
-
-    FAIL_IF_AFTER = ()
-
-    def __init__(self, independentVariable=None):
-        """
-        Constuctor.
-
-        Parameters
-        ----------
-        independentVariable : dict or None, optional
-            Name/value pairs to associate with the independent variable being modified
-            by this object.  Will be analyzed and plotted against other modifiers with
-            the same name.
-        """
-        if independentVariable is None:
-            independentVariable = {}
-        self.independentVariable = independentVariable
-
-    def __call__(self, cs, blueprints, geom):
-        """Perform the desired modifications to input objects."""
-        raise NotImplementedError
-
-
-class NeutronicMeshsSizeModifier(InputModifier):
-    """
-    Adjust the neutronics mesh in all assemblies by a multiplication factor.
-
-    This can be useful when switching between nodal and finite difference
-    approximations, or when doing mesh convergence sensitivity studies.
-
-    Attributes
-    ----------
-    multFactor : int
-        Factor to multiply the number of axial mesh points per block by.
-    """
-
-    def __init__(self, multFactor):
-        InputModifier.__init__(self, {self.__class__.__name__: multFactor})
-        if not isinstance(multFactor, int):
-            raise TypeError(
-                "multFactor must be an integer, but got {}".format(multFactor)
-            )
-        self.multFactor = multFactor
-
-    def __call__(self, cs, blueprints, geom):
-        for assemDesign in blueprints.assemDesigns:
-            assemDesign.axialMeshPoints = [
-                ax * self.multFactor for ax in assemDesign.axialMeshPoints
-            ]
-
-
-class FullCoreModifier(InputModifier):
-    """
-    Grow the SystemLayoutInput to from a symmetric core to a full core.
-    """
-
-    def __call__(self, cs, blueprints, geom):
-        geom.growToFullCore()
-
-
-class _PinTypeAssemblyModifier(InputModifier):
-    """
-    Abstract class for modifying something about a pin, within a block.
-
-    This will construct blocks, determine if the block should be modified by checking
-    the ``_getBlockTypesToModify``, and then run ``_adjustBlock(b)``. The ``Blueprints``
-    are then updated based on the modification assuming that dimension names match
-    exactly to ComponenBlueprint attributes (which is true, because ComponentBlueprint
-    attributes are programmatically derived from Component constructors).
-    """
-
-    def __init__(self, value):
-        InputModifier.__init__(self, {self.__class__.__name__: value})
-        self.value = value
-
-    def __call__(self, cs, blueprints, geom):
-        for bDesign in blueprints.blockDesigns:
-            # bDesign construct requires lots of arguments, many of which have no impact.
-            # The following can safely be defaulted to meaningless inputs:
-            # axialIndex: a block can be reused at any axial index, modifications made
-            #     dependent on will not translate back to the input in a  meaningful
-            #     fashion
-            # axialMeshPoints: similar to above, this is specified by the assembly, and
-            #     a block can be within any section of an assembly.
-            # height: similar to above. a block can have any height specified by an
-            #     assembly. if height-specific modifications are required, then a new
-            #     block definition should be created in the input
-            # xsType: similar to above. a block can have any xsType specified through
-            #     the assembly definition assembly. if xsType-specific modifications are
-            #     required, then a new block definition should be created in the input
-            # materialInput: this is the materialModifications from the assembly
-            #     definition. if material modifications are required on a block-specific
-            #     basis, they should be edited directly
-            b = bDesign.construct(
-                cs,
-                blueprints,
-                axialIndex=1,
-                axialMeshPoints=1,
-                height=1,
-                xsType="A",
-                materialInput={},
-            )
-
-            if not b.hasFlags(self._getBlockTypesToModify()):
-                continue
-
-            self._adjustBlock(b)
-
-            for cDesign, c in zip(bDesign, b):
-                for dimName in c.DIMENSION_NAMES:
-                    inpDim = getattr(cDesign, dimName)
-                    newDim = getattr(c.p, dimName)
-                    if isinstance(newDim, tuple):
-                        # map linked component dimension
-                        link = component._DimensionLink(newDim)
-                        newDim = str(link)
-                    if inpDim != newDim:
-                        setattr(cDesign, dimName, newDim)
-
-    def _getBlockTypesToModify(self):
-        """Hook method to determine blocks that should be modified."""
-        raise NotImplementedError
-
-    def _adjustBlock(self, b):
-        """Hook method for `__call__` template method."""
-        raise NotImplementedError
-
-
-class SmearDensityModifier(_PinTypeAssemblyModifier):
-    """
-    Adjust the smeared density to the specified value.
-
-    This is effectively how much of the space inside the cladding tube is occupied by
-    fuel at fabrication.
-
-    See Also
-    --------
-    armi.reactor.blocks.Block.adjustSmearDensity
-        Actually adjusts the smeared density
-    """
-
-    def _getBlockTypesToModify(self):
-        """Hook method to determine blocks that should be modified."""
-        return flags.Flags.FUEL
-
-    def _adjustBlock(self, b):
-        """Hook method for `__call__` template method."""
-        b.adjustSmearDensity(self.value)
-
-
-class CladThicknessByODModifier(_PinTypeAssemblyModifier):
-    """
-    Adjust the cladding thickness by adjusting the inner diameter of all cladding components.
-    """
-
-    FAIL_IF_AFTER = (SmearDensityModifier,)
-
-    def _getBlockTypesToModify(self):
-        """Hook method to determine blocks that should be modified."""
-        return ""
-
-    def _adjustBlock(self, b):
-        b.adjustCladThicknessByOD(self.value)
-
-
-class CladThicknessByIDModifier(_PinTypeAssemblyModifier):
-    """
-    Adjust the cladding thickness by adjusting the outer diameter of the cladding component.
-    """
-
-    FAIL_IF_AFTER = (SmearDensityModifier,)
-
-    def _getBlockTypesToModify(self):
-        """Hook method to determine blocks that should be modified."""
-        return ""
-
-    def _adjustBlock(self, b):
-        b.adjustCladThicknessByID(self.value)
-
-
-class NeutronicConvergenceModifier(InputModifier):
-    """
-    Adjust the neutronics convergence parameters ``epsEig``, ``epsFSAvg``, and ``epsFSPoint``.
-
-    The supplied value is used for ``epsEig``. ``epsFSAvg`` and ``epsFSPoint`` are set
-    to 100 times the supplied value.
-
-    This can be used to perform sensitivity studies on convergence criteria.
-    """
-
-    def __init__(self, value):
-        InputModifier.__init__(self, {self.__class__.__name__: value})
-        self.value = value
-        if value > 1e-2 or value <= 0.0:
-            raise ValueError(
-                "Neutronic convergence modifier value must be greater than 0 and less "
-                "than 1e-2 (got {})".format(value)
-            )
-
-    def __call__(self, cs, blueprints, geom):
-        cs["epsFSAvg"] = self.value * 100
-        cs["epsFSPoint"] = self.value * 100
-        cs["epsEig"] = self.value
-
-
-class SettingsModifier(InputModifier):
-    """
-    Adjust settings to specified values.
-    """
-
-    bad_setting_names = {
-        "epsEig": NeutronicConvergenceModifier,
-        "epsFSAvg": NeutronicConvergenceModifier,
-        "epsFSPoint": NeutronicConvergenceModifier,
-    }
-
-    def __init__(self, settingName, value):
-        InputModifier.__init__(self, independentVariable={settingName: value})
-        self.settingName = settingName
-        self.value = value
-        if settingName in self.bad_setting_names:
-            raise ValueError(
-                "An alternate modifier exists for adjusting the setting `{}`, use {}".format(
-                    settingName, self.bad_setting_names[settingName]
-                )
-            )
-
-    def __call__(self, cs, blueprints, geom):
-        cs[self.settingName] = self.value
